pyproject.toml
[tool.poetry]
name = "unique-merch"
version = "0.1.0"
description = ""
authors = ["AJ <akarsh.1995.02@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4"
pandas = "^1.2.2"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
